pyproject.toml
[project]
name = "cwtr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "black>=24.10.0",
    "bs4>=0.0.2",
    "cloudscraper>=1.2.71",
    "colorama>=0.4.6",
    "fake-useragent>=2.0.3",
    "openai>=1.59.8",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "requests>=2.32.3",
    "schedule>=1.2.2",
]

[tool.poetry.dependencies]
colorama = "^0.4.6"
